requirements.txt
streamlit==1.53.1
psycopg2-binary==2.9.9
